Refactor :  added a common index file for imports and exports
