+++ src/Context/index.js
@@ -1,0 +1,6 @@
+import { NoteProvider, useNote } from "./NoteContext";
+import { AuthProvider, useAuth } from "./AuthContext";
+import { DeleteProvider, useDelete } from "./DeleteContext";
+import { ArchiveProvider, useArchive } from "./ArchiveContext";
+
+export {NoteProvider, AuthProvider, DeleteProvider, ArchiveProvider, useNote, useAuth, useDelete, useArchive }
